Pass prepared finder to MutableIterator
