+++ src/MutableIterator.php
@@ -33,43 +33,12 @@ class MutableIterator implements \IteratorAggregate, \Countable
 
     /**
      * @param Container $container
-     * @param string[] $directories
-     * @param string[] $excludes
-     * @param string[] $names
+     * @param Finder $finder
      */
-    public function __construct(Container $container, $directories, $excludes, $names)
+    public function __construct(Container $container, Finder $finder)
     {
         $this->container = $container;
-        $this->finder = $this->prepareFinder($directories, $excludes, $names);
-    }
-
-    protected function prepareFinder($directories, $excludes, $names)
-    {
-        $finder = new Finder();
-
-        if (!is_null($names) && count($names) > 0) {
-            foreach ($names as $name) {
-                $finder->name($name);
-            }
-        } else {
-            $finder->name('*.php');
-        }
-
-        if ($directories) {
-            foreach ($directories as $directory) {
-                $finder->in($directory);
-            }
-        } else {
-            $finder->in('.');
-        }
-
-        if (isset($excludes)) {
-            foreach ($excludes as $exclude) {
-                $finder->exclude($exclude);
-            }
-        }
-
-        return $finder;
+        $this->finder = $finder;
     }
 
     /**
